Adds publication setup for bintray
https://bintray.com/openzipkin-contrib/maven/zipkin-secondary-sampling

First step post-merge is to check snapshots publish. After that, we
should do a release by creating a tag release-0.1.0

The first release will not complete because you cannot setup maven
central until there's a version. Once 0.1.0 exists in bintray, we can
click to allow it to be published.

See #22